rename docker2earth to docker2earthly
